requirements.txt
#Data manipulation and preprocessing
pandas==2.0.3

#Machine learning and deep learning
tensorflow==2.15.0
scikit-learn==1.3.0

#Plotting and visualizations
matplotlib==3.8.0

#Working with Google Cloud services
google-cloud-pubsub==2.14.1
google-cloud-bigquery==3.11.1